// apps/snsgov/src/candid/sns_governance.did.d.ts
export interface Account {
  'owner' : [] | [Principal],
  'subaccount' : [] | [Subaccount],
}
export type Action = {
    'ManageNervousSystemParameters' : NervousSystemParameters
  } |
  { 'AddGenericNervousSystemFunction' : NervousSystemFunction } |
  { 'ManageDappCanisterSettings' : ManageDappCanisterSettings } |
  { 'ExecuteExtensionOperation' : ExecuteExtensionOperation } |
  { 'UpgradeExtension' : UpgradeExtension } |
  { 'RemoveGenericNervousSystemFunction' : bigint } |
  { 'SetTopicsForCustomProposals' : SetTopicsForCustomProposals } |
  { 'RegisterExtension' : RegisterExtension } |
  { 'UpgradeSnsToNextVersion' : {} } |
  { 'RegisterDappCanisters' : RegisterDappCanisters } |
  { 'TransferSnsTreasuryFunds' : TransferSnsTreasuryFunds } |
  { 'UpgradeSnsControlledCanister' : UpgradeSnsControlledCanister } |
  { 'DeregisterDappCanisters' : DeregisterDappCanisters } |
  { 'MintSnsTokens' : MintSnsTokens } |
  { 'AdvanceSnsTargetVersion' : AdvanceSnsTargetVersion } |
  { 'Unspecified' : {} } |
  { 'ManageSnsMetadata' : ManageSnsMetadata } |
  {
    'ExecuteGenericNervousSystemFunction' : ExecuteGenericNervousSystemFunction
  } |
  { 'ManageLedgerParameters' : ManageLedgerParameters } |
  { 'Motion' : Motion };
export type ActionAuxiliary = {
    'TransferSnsTreasuryFunds' : MintSnsTokensActionAuxiliary
  } |
  { 'MintSnsTokens' : MintSnsTokensActionAuxiliary } |
  { 'AdvanceSnsTargetVersion' : AdvanceSnsTargetVersionActionAuxiliary };
export interface AddNeuronPermissions {
  'permissions_to_add' : [] | [NeuronPermissionList],
  'principal_id' : [] | [Principal],
}
export interface AdvanceSnsTargetVersion { 'new_target' : [] | [SnsVersion] }
export interface AdvanceSnsTargetVersionActionAuxiliary {
  'target_version' : [] | [SnsVersion],
}
export interface Amount { 'e8s' : bigint }
export interface Ballot {
  'vote' : number,
  'cast_timestamp_seconds' : bigint,
  'voting_power' : bigint,
}
// ...
export interface CustomProposalCriticality {
  'additional_critical_native_action_ids' : BigUint64Array,
}
export interface Decimal { 'human_readable' : [] | [string] }
export interface DefaultFollowees { 'followees' : Array<[bigint, Followees]> }
// ...
export interface DisburseMaturityInProgress {
  'timestamp_of_disbursement_seconds' : bigint,
  'amount_e8s' : bigint,
  'account_to_disburse_to' : [] | [Account],
  'finalize_disbursement_timestamp_seconds' : [] | [bigint],
}
export interface DisburseMaturityResponse {
  'amount_disbursed_e8s' : bigint,
  'amount_deducted_e8s' : [] | [bigint],
}
export interface DisburseResponse { 'transfer_block_height' : bigint }
export type DissolveState = { 'DissolveDelaySeconds' : bigint } |
  { 'WhenDissolvedTimestampSeconds' : bigint };
// ...
export interface Followee {
  'alias' : [] | [string],
  'neuron_id' : [] | [NeuronId],
}
export interface Followees { 'followees' : Array<NeuronId> }
export interface FolloweesForTopic {
  'topic' : [] | [Topic],
  'followees' : Array<Followee>,
}
// ...
export interface Governance {
  'root_canister_id' : [] | [Principal],
  'timers' : [] | [Timers],
  'cached_upgrade_steps' : [] | [CachedUpgradeSteps],
  'id_to_nervous_system_functions' : Array<[bigint, NervousSystemFunction]>,
  'metrics' : [] | [GovernanceCachedMetrics],
  'maturity_modulation' : [] | [MaturityModulation],
  'upgrade_journal' : [] | [UpgradeJournal],
  'mode' : number,
  'parameters' : [] | [NervousSystemParameters],
  'is_finalizing_disburse_maturity' : [] | [boolean],
  'deployed_version' : [] | [Version],
  'sns_initialization_parameters' : string,
  'latest_reward_event' : [] | [RewardEvent],
  'pending_version' : [] | [PendingVersion],
  'swap_canister_id' : [] | [Principal],
  'ledger_canister_id' : [] | [Principal],
  'proposals' : Array<[bigint, ProposalData]>,
  'in_flight_commands' : Array<[string, NeuronInFlightCommand]>,
  'sns_metadata' : [] | [ManageSnsMetadata],
  'neurons' : Array<[string, Neuron]>,
  'target_version' : [] | [Version],
  'genesis_timestamp_seconds' : bigint,
}
// ...
export interface GovernanceError {
  'error_message' : string,
  'error_type' : number,
}
// ...
export interface MintSnsTokens {
  'to_principal' : [] | [Principal],
  'to_subaccount' : [] | [Subaccount],
  'memo' : [] | [bigint],
  'amount_e8s' : [] | [bigint],
}
export interface MintSnsTokensActionAuxiliary { 'valuation' : [] | [Valuation] }
export interface Motion { 'motion_text' : string }
export interface NervousSystemFunction {
  'id' : bigint,
  'name' : string,
  'description' : [] | [string],
  'function_type' : [] | [FunctionType],
}
export interface NervousSystemParameters {
  'default_followees' : [] | [DefaultFollowees],
  'max_dissolve_delay_seconds' : [] | [bigint],
  'max_dissolve_delay_bonus_percentage' : [] | [bigint],
  'max_followees_per_function' : [] | [bigint],
  'automatically_advance_target_version' : [] | [boolean],
  'neuron_claimer_permissions' : [] | [NeuronPermissionList],
  'neuron_minimum_stake_e8s' : [] | [bigint],
  'max_neuron_age_for_age_bonus' : [] | [bigint],
  'initial_voting_period_seconds' : [] | [bigint],
  'neuron_minimum_dissolve_delay_to_vote_seconds' : [] | [bigint],
  'reject_cost_e8s' : [] | [bigint],
  'max_proposals_to_keep_per_action' : [] | [number],
  'wait_for_quiet_deadline_increase_seconds' : [] | [bigint],
  'max_number_of_neurons' : [] | [bigint],
  'transaction_fee_e8s' : [] | [bigint],
  'custom_proposal_criticality' : [] | [CustomProposalCriticality],
  'max_number_of_proposals_with_ballots' : [] | [bigint],
  'max_age_bonus_percentage' : [] | [bigint],
  'neuron_grantable_permissions' : [] | [NeuronPermissionList],
  'voting_rewards_parameters' : [] | [VotingRewardsParameters],
  'maturity_modulation_disabled' : [] | [boolean],
  'max_number_of_principals_per_neuron' : [] | [bigint],
}
export interface Neuron {
  'id' : [] | [NeuronId],
  'staked_maturity_e8s_equivalent' : [] | [bigint],
  'permissions' : Array<NeuronPermission>,
  'maturity_e8s_equivalent' : bigint,
  'cached_neuron_stake_e8s' : bigint,
  'created_timestamp_seconds' : bigint,
  'topic_followees' : [] | [
    { 'topic_id_to_followees' : Array<[number, FolloweesForTopic]> }
  ],
  'source_nns_neuron_id' : [] | [bigint],
  'auto_stake_maturity' : [] | [boolean],
  'aging_since_timestamp_seconds' : bigint,
  'dissolve_state' : [] | [DissolveState],
  'voting_power_percentage_multiplier' : bigint,
  'vesting_period_seconds' : [] | [bigint],
  'disburse_maturity_in_progress' : Array<DisburseMaturityInProgress>,
  'followees' : Array<[bigint, Followees]>,
  'neuron_fees_e8s' : bigint,
}
export interface NeuronId { 'id' : Uint8Array }
export interface NeuronIds { 'neuron_ids' : Array<NeuronId> }
export interface NeuronInFlightCommand {
  'command' : [] | [Command_2],
  'timestamp' : bigint,
}
export interface NeuronPermission {
  'principal' : [] | [Principal],
  'permission_type' : Int32Array,
}
export interface NeuronPermissionList { 'permissions' : Int32Array }
// ...
export interface Percentage { 'basis_points' : [] | [bigint] }
// ...
export interface Proposal {
  'url' : string,
  'title' : string,
  'action' : [] | [Action],
  'summary' : string,
}
export interface ProposalData {
  'id' : [] | [ProposalId],
  'payload_text_rendering' : [] | [string],
  'topic' : [] | [Topic],
  'action' : bigint,
  'failure_reason' : [] | [GovernanceError],
  'action_auxiliary' : [] | [ActionAuxiliary],
  'ballots' : Array<[string, Ballot]>,
  'minimum_yes_proportion_of_total' : [] | [Percentage],
  'reward_event_round' : bigint,
  'failed_timestamp_seconds' : bigint,
  'reward_event_end_timestamp_seconds' : [] | [bigint],
  'proposal_creation_timestamp_seconds' : bigint,
  'initial_voting_period_seconds' : bigint,
  'reject_cost_e8s' : bigint,
  'latest_tally' : [] | [Tally],
  'wait_for_quiet_deadline_increase_seconds' : bigint,
  'decided_timestamp_seconds' : bigint,
  'proposal' : [] | [Proposal],
  'proposer' : [] | [NeuronId],
  'wait_for_quiet_state' : [] | [WaitForQuietState],
  'minimum_yes_proportion_of_exercised' : [] | [Percentage],
  'is_eligible_for_rewards' : boolean,
  'executed_timestamp_seconds' : bigint,
}
export interface ProposalId { 'id' : bigint }
// ...
export interface Tally {
  'no' : bigint,
  'yes' : bigint,
  'total' : bigint,
  'timestamp_seconds' : bigint,
}
// ...
export type Topic = { 'DappCanisterManagement' : null } |
  { 'DaoCommunitySettings' : null } |
  { 'ApplicationBusinessLogic' : null } |
  { 'CriticalDappOperations' : null } |
  { 'TreasuryAssetManagement' : null } |
  { 'Governance' : null } |
  { 'SnsFrameworkManagement' : null };
// ...
export interface TransferSnsTreasuryFunds {
  'from_treasury' : number,
  'to_principal' : [] | [Principal],
  'to_subaccount' : [] | [Subaccount],
  'memo' : [] | [bigint],
  'amount_e8s' : bigint,
}
// ...
export interface VotingRewardsParameters {
  'final_reward_rate_basis_points' : [] | [bigint],
  'initial_reward_rate_basis_points' : [] | [bigint],
  'reward_rate_transition_duration_seconds' : [] | [bigint],
  'round_duration_seconds' : [] | [bigint],
}
export interface WaitForQuietState {
  'current_deadline_timestamp_seconds' : bigint,
}